store uid and phonbeNumber
